# rust: Use vectorized AES-NI to implement AES-CTR