--- a/rust/src/internals/mod.rs
+++ b/rust/src/internals/mod.rs
@@ -8,7 +8,7 @@ mod ctr;
 mod xor;
 
 pub use self::aes::{Aes128, Aes256};
-pub use self::block::Block;
+pub use self::block::{Block, Block8};
 pub use self::block::SIZE as BLOCK_SIZE;
 pub use self::block_cipher::BlockCipher;
 pub use self::cmac::Cmac;