Adds a `Block8` type for performing vectorized AES operations. Uses `Block8` to implement the internals of AES-CTR.

Provides an approximately 1.4X speedup (Unfortunately AES-CMAC can't be vectorized and becomes the bottleneck).

Before:

```
test bench::bench_aes_siv_128_encrypt_16384_bytes ... bench:
51,337 ns/iter (+/- 6,315) = 319 MB/s
```

After:

```
test bench::bench_aes_siv_128_encrypt_16384_bytes ... bench:
39,036 ns/iter (+/- 5,027) = 419 MB/s
```